+++ packages/runtime/adapters/eventstore-adapters/eventstore-postgresql-serverless/src/escape.ts
@@ -1,4 +1,0 @@
-const escape = (str: any): string =>
-  `'${String(str).replace(/(['])/gi, '$1$1')}'`
-
-export default escape