+++ tests/readmodel-test-utils/index.ts
@@ -1,12 +1,6 @@
-import * as AWS from 'aws-sdk'
 import createSqliteAdapter from '@resolve-js/readmodel-lite'
-import createPostgresqlServerlessAdapter from '@resolve-js/readmodel-postgresql-serverless'
 import createPostgresqlAdapter from '@resolve-js/readmodel-postgresql'
 import createMySQLAdapter from '@resolve-js/readmodel-mysql'
-import {
-  create as createPostgresServerlessResource,
-  destroy as destroyPostgresServerlessResource,
-} from '@resolve-js/readmodel-postgresql-serverless'
 import {
   create as createPostgresResource,
   destroy as destroyPostgresResource,
@@ -15,35 +9,6 @@ import {
   create as createMySQLResource,
   destroy as destroyMySQLResource,
 } from '@resolve-js/readmodel-mysql'
-
-export function isPostgresServerless(): boolean {
-  if (
-    process.env.TEST_POSTGRES_SERVERLESS !== undefined &&
-    process.env.TEST_POSTGRES_SERVERLESS !== 'false'
-  ) {
-    if (process.env.AWS_ACCESS_KEY_ID == null) {
-      throw new Error(`Environment variable AWS_ACCESS_KEY_ID is required`)
-    }
-    if (process.env.AWS_SECRET_ACCESS_KEY == null) {
-      throw new Error(`Environment variable AWS_SECRET_ACCESS_KEY is required`)
-    }
-    if (process.env.AWS_RDS_CLUSTER_ARN == null) {
-      throw new Error(`Environment variable AWS_RDS_CLUSTER_ARN is required`)
-    }
-    if (process.env.AWS_RDS_ADMIN_SECRET_ARN == null) {
-      throw new Error(
-        `Environment variable AWS_RDS_ADMIN_SECRET_ARN is required`
-      )
-    }
-    AWS.config.update({
-      credentials: {
-        accessKeyId: process.env.AWS_ACCESS_KEY_ID,
-        secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
-      },
-    })
-    return true
-  }
-}
 
 export function isPostgres(): boolean {
   if (
@@ -103,8 +68,6 @@ export function jestTimeout(): number {
 
 let adapters: Record<string, any> = {}
 
-const uniquePostfix = `${process.pid}_${Math.round(Math.random() * 1000)}`
-
 const proxy = new Proxy(
   {},
   {
@@ -133,16 +96,6 @@ const proxy = new Proxy(
 
 export { proxy as adapters }
 
-export function getPostgresServerlessOptions(uniqueName: string) {
-  return {
-    databaseName: `${uniqueName}_${uniquePostfix}`,
-    dbClusterOrInstanceArn: process.env.AWS_RDS_CLUSTER_ARN,
-    awsSecretStoreAdminArn: process.env.AWS_RDS_ADMIN_SECRET_ARN,
-    region: process.env.AWS_REGION ?? 'eu-central-1',
-    userLogin: process.env.AWS_RDS_USERNAME ?? 'master',
-  }
-}
-
 export function getPostgresOptions(uniqueName: string) {
   return {
     databaseName: uniqueName,
@@ -165,39 +118,7 @@ export function getMySQLOptions(uniqueName: string) {
   }
 }
 
-export const adapterFactory = isPostgresServerless()
-  ? {
-      name: '@resolve-js/readmodel-postgresql-serverless',
-      create(uniqueName: string) {
-        return async () => {
-          const options = getPostgresServerlessOptions(uniqueName)
-
-          try {
-            await destroyPostgresServerlessResource(options)
-          } catch {}
-          await createPostgresServerlessResource(options)
-
-          adapters[uniqueName] = createPostgresqlServerlessAdapter({
-            databaseName: options.databaseName,
-            dbClusterOrInstanceArn: options.dbClusterOrInstanceArn,
-            awsSecretStoreArn: options.awsSecretStoreAdminArn,
-            region: options.region,
-          })
-        }
-      },
-      destroy(uniqueName: string) {
-        return async () => {
-          const options = getPostgresServerlessOptions(uniqueName)
-
-          await adapters[uniqueName].dispose()
-
-          await destroyPostgresServerlessResource(options)
-
-          delete adapters[uniqueName]
-        }
-      },
-    }
-  : isPostgres()
+export const adapterFactory = isPostgres()
   ? {
       name: '@resolve-js/readmodel-postgresql',
       create(uniqueName: string) {